# Documentazione/Progettazione/openapi.yaml
openapi: 3.0.3
info:
  title: Monopoly IoT API
  version: 1.0.0
  description: |
    API REST esposte dall'API Gateway del sistema Monopoly IoT.

    Il sistema è composto da 5 microservizi:
    - **api-gateway** (porta 8080) — punto di ingresso unico
    - **user-service** (porta 8081) — gestione utenti
    - **match-service** (porta 8082) — partite, tornei, motore di gioco
    - **board-service** (porta 8083) — sedi, tavoli, zone RFID
    - **iot-gateway** (porta 8085) — acquisizione eventi MQTT, buffer offline

    Tutte le richieste vengono inoltrate dal Gateway ai servizi interni.
servers:
  - url: http://localhost:8080
    description: API Gateway locale
...
paths:
...
  /api/matches/create:
    post:
      tags: [Matches]
      summary: Crea partita pianificata (status=CREATED)
      operationId: createMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gameTableId]
              properties:
                gameTableId:
                  type: integer
                  format: int64
                  example: 1
                maxPlayers:
                  type: integer
                  default: 4
                  example: 4
                scheduledTime:
                  type: string
                  example: "2026-06-10T18:00:00"
      responses:
        "200":
          description: Partita creata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: gameTableId mancante
  /api/matches/start:
    post:
      tags: [Matches]
      summary: Crea e avvia direttamente una partita (status=STARTED)
      operationId: startMatchDirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Match"
      responses:
        "200":
          description: Partita avviata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /api/matches/join:
    post:
      tags: [Matches]
      summary: Iscrive un giocatore a una partita
      description: |
        Se `matchId` è specificato, iscrive il giocatore a quella partita.
        Altrimenti cerca la prima partita in stato CREATED, o ne crea una nuova.
      operationId: joinMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
                  format: int64
                  example: 42
                matchId:
                  type: integer
                  format: int64
                  example: 7
                  description: Opzionale. Se omesso, viene usato matchmaking automatico.
      responses:
        "200":
          description: Iscrizione completata
          content:
            application/json:
              schema:
                type: string
                example: "Iscritto con successo alla Partita #7"
        "400":
          description: Partita piena, già iscritto o stato non valido
  /api/matches/leave:
    post:
      tags: [Matches]
      summary: Annulla iscrizione a una partita (solo se status=CREATED)
      operationId: leaveMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, matchId]
              properties:
                userId:
                  type: integer
                  format: int64
                matchId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Iscrizione annullata
        "400":
          description: Non iscritto o partita già avviata
  /api/matches/{id}/start-match:
    post:
      tags: [Matches]
      summary: Avvia una partita esistente (CREATED → STARTED)
      description: |
        Imposta status=STARTED, startTime e currentTurnPlayerId (giocatore con playerOrder=1).
        Solo partite in stato CREATED o PAUSED possono essere avviate.
      operationId: startMatchById
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Partita avviata
          content:
            application/json:
              schema:
                type: string
                example: "Partita #5 avviata con successo!"
        "400":
          description: Stato partita non valido per l'avvio
        "404":
          description: Partita non trovata
  /api/matches/{id}/pause:
    post:
      tags: [Matches]
      summary: Mette in pausa una partita (STARTED → PAUSED)
      operationId: pauseMatch
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Partita in pausa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Partita non trovata
  /api/matches/{id}/end:
    post:
      tags: [Matches]
      summary: Termina una partita (→ FINISHED)
      description: |
        Calcola il vincitore per patrimonio netto (saldo + valore proprietà).
        Aggiorna le statistiche di giocatori, tavoli e sede.
      operationId: endMatch
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Partita terminata con vincitore calcolato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Partita non trovata
  /api/matches/{id}:
    delete:
      tags: [Matches]
      summary: Elimina partita e tutti i dati associati
      description: Elimina anche SensorEvent, MatchPropertyState e MatchPlayer relativi.
      operationId: deleteMatch
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Partita eliminata
        "404":
          description: Partita non trovata
  /api/matches/{id}/schedule:
    put:
      tags: [Matches]
      summary: Aggiorna pianificazione di una partita
      operationId: scheduleMatch
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledTime:
                  type: string
                  example: "2026-06-15T20:00:00"
                gameTableId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Pianificazione aggiornata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /api/matches/{id}/players:
    get:
      tags: [Matches]
      summary: Elenco giocatori della partita
      operationId: getMatchPlayers
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lista MatchPlayer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchPlayer"
  /api/matches/{id}/properties:
    get:
      tags: [Matches]
      summary: Stato proprietà della partita
      operationId: getMatchProperties
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lista MatchPropertyState
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchPropertyState"
  /api/matches/{id}/events:
    get:
      tags: [Matches]
      summary: Ultimi eventi sensore della partita (max 1000)
      operationId: getMatchEvents
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lista SensorEvent ordinati per id decrescente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorEvent"
  /api/matches/{id}/rotate-turn:
    post:
      tags: [Matches]
      summary: Passa il turno al prossimo giocatore non fallito
      operationId: rotateTurn
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Turno aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Partita non trovata
  /api/matches/{id}/reset:
    post:
      tags: [Matches]
      summary: Reset partita (riparte da STARTED, saldo 1500 per tutti)
      operationId: resetMatch
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Reset completato
          content:
            application/json:
              schema:
                type: string
                example: "Reset completed successfully for Match #3!"
  /api/matches/simulate-event:
    post:
      tags: [Matches]
      summary: Processa evento sensore nel motore di gioco
      description: |
        Endpoint chiamato dall'IoT Gateway (e usabile per test manuali).
        Salva l'evento in match.db e lo processa nel GameEngineService.
        La partita deve essere in stato STARTED.
      operationId: simulateSensorEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorEvent"
            example:
              matchId: 1
              zoneId: 12
              taggedObjectId: 101
              eventType: "ENTER"
              rawPayload: "{\"source\":\"rfid-reader-3\"}"
      responses:
        "200":
          description: Evento salvato e processato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorEvent"
        "400":
          description: matchId mancante o partita non in stato STARTED
        "500":
          description: Errore nel motore di gioco

  # ─────────────────────────────────────────────────────────────
  # TOURNAMENTS (→ match-service :8082)
  # ─────────────────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Identificatore univoco della risorsa
  schemas:
...
    Match:
      type: object
      description: Partita di Monopoly IoT
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        gameTableId:
          type: integer
          format: int64
        status:
          type: string
          enum: [CREATED, STARTED, PAUSED, FINISHED, CANCELLED, SCHEDULED]
          example: "CREATED"
        currentTurnPlayerId:
          type: integer
          format: int64
          description: ID del MatchPlayer a cui spetta il turno corrente
          nullable: true
        scheduledTime:
          type: string
          example: "2026-06-10T18:00:00"
        maxPlayers:
          type: integer
          default: 4
        tournamentId:
          type: integer
          format: int64
          nullable: true
        winnerPlayerId:
          type: integer
          format: int64
          nullable: true
          description: userId del vincitore (impostato al termine)
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
    MatchPlayer:
      type: object
      description: Giocatore iscritto a una specifica partita
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        matchId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        playerOrder:
          type: integer
          description: Ordine di turno (1 = primo)
        currentBalance:
          type: integer
          description: Saldo attuale in lire Monopoly (iniziale = initialBalance da GameSettings)
          example: 1500
        inGame:
          type: integer
          description: 1 = attivo in partita, 0 = eliminato/uscito
          enum: [0, 1]
        bankrupt:
          type: integer
          description: 1 = fallito, 0 = ancora in gioco
          enum: [0, 1]
    MatchPropertyState:
      type: object
      description: Stato di una proprietà Monopoly durante una partita
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        matchId:
          type: integer
          format: int64
        propertyId:
          type: integer
          format: int64
          description: Riferimento a monopoly_properties.id
        ownerMatchPlayerId:
          type: integer
          format: int64
          nullable: true
          description: ID del MatchPlayer proprietario (null = non acquistata)
        mortgaged:
          type: integer
          enum: [0, 1]
          description: 1 = ipotecata
        houses:
          type: integer
          description: Numero di case costruite (0-4)
        hotels:
          type: integer
          description: Numero di alberghi (0-1)
...